+++ components/pm-subscription-handler/pmsh_service/mod/api/controller.py
@@ -76,3 +76,33 @@ def post_subscription(body):
                      exc_info=True)
         response = e.invalid_message, HTTPStatus.BAD_REQUEST.value
     return response
+
+
+def get_subscription_by_name(subscription_name):
+    """
+    Retrieves subscription based on the input subscription Name
+
+    Args:
+        subscription_name(String): Name of the subscription to fetch subscription from DB
+    Returns:
+       SubscriptionModel: Single Subscription information as JSON
+    Raises:
+        Error: If subscription is not defined (or) Exception occurs while retrieving
+               from DB
+    """
+    try:
+        subscription = subscription_service.get_subscription_by_name(subscription_name)
+        if subscription is not None:
+            logger.info(f'subscription object with the name "{subscription_name}" '
+                        f'is fetched successfully from DB')
+            return subscription_service.subscription_encoder(subscription), HTTPStatus.OK
+        else:
+            logger.error(f'subscription object with the name "{subscription_name}" '
+                         f'is un successful to fetch from DB')
+            return {'error': f'Subscription is not defined with the name : '
+                             f'{subscription_name}'}, HTTPStatus.NOT_FOUND
+    except Exception as exception:
+        logger.error(f'The following exception occurs "{exception}" while fetching subscription '
+                     f'with the name "{subscription_name}"')
+        return {'error': f'Request is not processed due to Exception : '
+                         f'{exception}'}, HTTPStatus.INTERNAL_SERVER_ERROR